feat: Refactor flask structure
Follow best practices for scalability using Blueprints, app factory and other flask features
