+++ Dockerfile
@@ -9,11 +9,10 @@ WORKDIR /home/pokeapi
 COPY requirements.txt requirements.txt
 RUN python -m venv venv
 RUN venv/bin/pip install -r requirements.txt
-RUN venv/bin/pip install gunicorn
 
-COPY pokeapi.py boot.sh ./
-COPY utils/ ./utils
+COPY . .
 RUN chmod +x boot.sh
+RUN chmod +x test.sh
 
 ENV FLASK_APP pokeapi.py
 ENV POKEAPI_BERRY_BASE_URL=$ARG_POKEAPI_BERRY_BASE_URL
