+++ config.py
@@ -1,0 +1,5 @@
+import os
+
+class Config(object):
+    # Default config values
+    pass
